Move agent2 up to top level
And implement shell agent with tests.
